Docker setup (#1428)
* add metis config

* make add to wallet configurable per network

* fix add network to wallet

* configure metis token

* properly invalidate after creating new drip list

* wip

* add railway-specific dockerfile

* delete outdated dev dockerfile

* fix codegen

* remove Railway-specific workarounbds from main Dockerfile

* fix generate-schema script

* fix generate-schema script

* fix railway build

* adjust comments

* maybe this will make it work

* clean apt in railway build

* try this

* i don't know anymore

* how bout this

* for fucks sake!!!!!!!!

* that wasn't it

* restore package-lock from main

* back to this

* add publish-docker workflow

* add env vars back for railway

* fix script path

* fix generate schema script

* please?

* how about now

* god fucking damnit

* ...

* fix

* fix

* pls

* use metadata-action

* fixes

* do a multi-platform build

* make it run on staging

* change name of action

* add buildx setup

* add QEMU emulator

* only build on arm, looks like not enough memory on runner

* remove necessity for RAILWAY_DOCKERFILE_PATH var on railway

* prepare full local env

* make fiat estimates work for METIS

* work

* auto remove repo owner update container

* fix needing schema.graphql for local env

* add test token as default token

* add multiplayer

* fix missing PUBLIC_IPFS_GATEWAY_URL

* update docs

* add instructions for running different tags

* fix unable to pin sth

* try this

* try this

* now?????

* fix contracts tag env var name, fix add network to wallet

* change contracts default tag

* add npm i instruction to DEVELOPMENT.md to make commit linting work

* add section on running without Docker

* remove Metis for now

* set default tag for testnet service

* improve dev docs

* remove docker publish of the app container for now

* add search to local env

* remove metis chain ID

* remove leftovers from old E2E setup, fix tests

* add cache invalidation

* improve docs

* fix railway deploy

* actually fix railway build

* fix gh actions

* delete accidentally committed file

* fix unit tests

* actually do that

* again

* fix svelte check

* fix svelte check

+++ scripts/build-gql-types.sh
@@ -1,12 +1,21 @@
 #!/bin/bash
 
+echo "🦺 Building GraphQL types based on ./schema.graphql..."
+
+# If ./src/lib/graphql/__generated__/schema.graphql file doesn't exist, print an error and exit with 1
+if [ ! -f ./schema.graphql ]; then
+  echo "Error: ./src/lib/graphql/__generated__/schema.graphql file doesn't exist. Run 'npm run generate-schema' first."
+  exit 1
+fi
+
+
 start_time=$(date +%s)
 end_time=$((start_time + 300))  # 5 minutes = 300 seconds
 
 find . -regex './src/.*__generated__.*' -type d -prune -exec rm -r "{}" \;
 
 while [[ $(date +%s) -lt $end_time ]]; do
-  graphql-codegen
+  npx graphql-codegen
   if [[ $? -eq 0 ]]; then
     exit 0  # Success
   fi
